Feat: implemented stripe payment

+++ store/models.py
@@ -3,8 +3,8 @@ from django.db import models
 
 class Item(models.Model):
     name = models.CharField(max_length=200)
-    description = models.TextField(blank=True, null=True)
-    price = models.PositiveIntegerField()
+    description = models.TextField(blank=True, null=True)   # optional field
+    price = models.DecimalField(max_digits=10, decimal_places=2)
 
     def __str__(self):
         return f"{self.name} | {self.price}"
